Adding types of users so that we differentiate between vendors and clients, and also create suscriptions.

+++ bookings/src/main/java/com/kike/events/bookings/service/impl/BookingServiceImpl.java
@@ -5,17 +5,22 @@ import com.kike.events.bookings.dto.ResponseDto;
 import com.kike.events.bookings.entity.Booking;
 import com.kike.events.bookings.exception.BookingAlreadyExistsException;
 import com.kike.events.bookings.exception.ResourceNotFoundException;
+import com.kike.events.bookings.exception.UnauthorizedException;
 import com.kike.events.bookings.mapper.BookingMapper;
 import com.kike.events.bookings.repository.BookingRepository;
 import com.kike.events.bookings.service.IBookingService;
+import com.kike.events.bookings.service.auth.JwtService;
 import com.kike.events.bookings.service.client.EventsFeignClient;
 import lombok.AllArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
 import java.util.Optional;
 
 
@@ -25,22 +30,28 @@ public class BookingServiceImpl implements IBookingService {
 
     private BookingRepository bookingRepository;
     private EventsFeignClient eventsFeignClient;
+    private JwtService jwtService;
     private static final Logger log = LoggerFactory.getLogger(BookingServiceImpl.class);
 
 
     @Override
+    public void createBooking(BookingDto bookingDto, String userId, Jwt jwt) {
 
-    public void createBooking(BookingDto bookingDto, String userId) {
+        if(jwt == null)
+            throw new UnauthorizedException("Oauth2 access token is required");
 
-        Optional<Booking> existingBooking = bookingRepository.findByUserIdAndEventId(userId, bookingDto.getEventId());
+        String actualUserId = jwtService.getRealmRoles(jwt).contains("admin") ?
+                userId : jwtService.getUserId(jwt);
+
+        Optional<Booking> existingBooking = bookingRepository.findByUserIdAndEventId(actualUserId, bookingDto.getEventId());
 
         if (existingBooking.isPresent())
             throw new BookingAlreadyExistsException("The book with serviceId: " +
                     bookingDto.getEventId() + " and userId: " +
-                    userId + " already exists");
+                    actualUserId + " already exists");
 
         Booking booking = BookingMapper.mapToBooking(bookingDto, new Booking());
-        booking.setUserId(userId);
+        booking.setUserId(actualUserId);
 
         ResponseEntity<ResponseDto> responseDtoResponseEntity = eventsFeignClient.updateCurrentBookingsCount(bookingDto.getEventId());
 
@@ -59,11 +70,16 @@ public class BookingServiceImpl implements IBookingService {
     }
 
     @Override
-    public boolean updatebooking(BookingDto bookingDto, String userId) {
+    public boolean updatebooking(BookingDto bookingDto, String userId, Jwt jwt) {
 
+        if(jwt == null)
+            throw new UnauthorizedException("Oauth2 access token is required");
+
+        String actualUserId = jwtService.getRealmRoles(jwt).contains("admin") ?
+                userId : jwtService.getUserId(jwt);
 
         Booking booking = bookingRepository.findByUserIdAndEventId(userId, bookingDto.getEventId()).orElseThrow(() ->
-                new ResourceNotFoundException("Booking", "userId or eventId", userId + " " + bookingDto.getEventId() + " respectively"));
+                new ResourceNotFoundException("Booking", "userId or eventId", actualUserId + " " + bookingDto.getEventId() + " respectively"));
 
 
         BookingMapper.mapToBooking(bookingDto, booking);
@@ -75,11 +91,28 @@ public class BookingServiceImpl implements IBookingService {
     }
 
     @Override
-    public void deletebooking(Long eventId, String userId) {
+    public void deleteBooking(Long eventId, String userId, Jwt jwt) {
+
+
+        if(jwt == null)
+            throw new UnauthorizedException("Oauth2 access token is required");
+
+        // If the user is admin, then it chooses the userId, if not it's own userId is taken
+        String actualUserId = jwtService.getRealmRoles(jwt).contains("admin") ?
+                userId : jwtService.getUserId(jwt);
+
         Booking booking = bookingRepository.findByUserIdAndEventId(userId, eventId).orElseThrow(() ->
-                new ResourceNotFoundException("Booking", "userId or eventId", userId + " " + eventId + " respectively"));
+                new ResourceNotFoundException("Booking", "userId or eventId", actualUserId + " " + eventId + " respectively"));
 
         bookingRepository.deleteById(booking.getId());
+    }
+
+    @Override
+    @Transactional
+    public void deleteAllbookingsWithEventId(Long eventId) {
+        List<Booking> bookings = bookingRepository.findByEventId(eventId);
+
+        bookings.forEach(booking -> bookingRepository.deleteById(booking.getId()));
     }
 
 //    @Override
